% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_statgarten.R
\name{theme_statgarten}
\alias{theme_statgarten}
\title{Add Statgarten theme to ggplot chart}
\usage{
theme_statgarten(base_family = "NanumGothic")
}
\description{
This function allows you to add the statgarten theme to your ggplot graphics.
}
\examples{
library(dplyr)
library(ggplot2)
library(datatoys)
datatoys::birthRate \%>\%
  group_by(시점) \%>\%
  summarise(합계출산율 = mean(합계출산율, na.rm = TRUE)) \%>\%
  ggplot(aes(시점, 합계출산율)) +
  geom_point() +
  geom_line() +
  theme_statgarten()
}
\keyword{theme_statgarten}
